+++ src/imio/smartweb/common/behaviors/topics.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
+from imio.smartweb.common.widgets.select import TranslatedAjaxSelectFieldWidget
 from imio.smartweb.locales import SmartwebMessageFactory as _
-from plone.app.z3cform.widget import AjaxSelectFieldWidget
 from plone.autoform import directives
 from plone.autoform.interfaces import IFormFieldProvider
 from plone.supermodel import model
@@ -21,4 +21,4 @@ class ITopics(model.Schema):
         default=[],
         required=False,
     )
-    directives.widget(topics=AjaxSelectFieldWidget)
+    directives.widget(topics=TranslatedAjaxSelectFieldWidget)